move compression to chunks
